Fix some pedantic issues in GCC 13.3
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6935 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ rephrase/document.cc
@@ -1314,7 +1314,7 @@ Document::PackBoxes(Algorithm algo,
       for (const std::vector<BoxesAndGlue::BoxOut> &line : lines) {
         for (const BoxOut &out : line) {
           const int idx = out.box_idx;
-          CHECK(idx >= 0 && idx < boxes.size());
+          CHECK(idx >= 0 && idx < (int)boxes.size());
           // const BoxIn &in = boxes[idx];
           double glue = out.actual_glue;
